first part 100% done

+++ routes/web.php
@@ -37,23 +37,26 @@ Route::get('/', [ArticlesController::class,'index'])->name('articles');
 Route::get('/editarticles/{id}',[ArticlesController::class,'edit'])->name('editarticles');
 Route::PUT('/{id}/updatearticle',[ArticlesController::class,'update']);
 Route::DELETE('/{id}/deletearticle',[ArticlesController::class,'destroy']);
+
 //users
-Route::get('/users', [RegisteredUserController::class,'index'])->name('users');
+Route::get('/users', [RegisteredUserController::class,'index'])->middleware(['adminrole'])->name('users');
+Route::get('/createuser', [RegisteredUserController::class,'createuser'])->middleware(['adminrole'])->name('createuser');
+Route::post('/storeeuser', [RegisteredUserController::class,'storeuser'])->middleware(['adminrole']);
 Route::get('/edituser/{id}',[RegisteredUserController::class,'edit'])->middleware(['adminrole']);
-Route::PUT('/{id}/updateuser',[RegisteredUserController::class,'update']);
-Route::DELETE('/{id}/deleteuser',[RegisteredUserController::class,'destroy']);
+Route::PUT('/{id}/updateuser',[RegisteredUserController::class,'update'])->middleware(['adminrole']);
+Route::DELETE('/{id}/deleteuser',[RegisteredUserController::class,'destroy'])->middleware(['adminrole']);
 
 
 
 
 //avatars
-Route::get('/avatars', [AvatarController::class,'index'])->name('avatars');
-Route::get('/createavatar', [AvatarController::class,'create'])->name('createavatar');
-Route::post('/storeavatar', [AvatarController::class,'store']);
-Route::get('/showavatar', [AvatarController::class,'show'])->name('showavatar');
-Route::get('/editavatar/{id}', [AvatarController::class,'edit'])->name('editavatar');
-Route::put('/{id}/updateavatar', [AvatarController::class,'update']);
-Route::delete('/{id}/deleteavatar', [AvatarController::class,'destroy']);
+Route::get('/avatars', [AvatarController::class,'index'])->name('avatars')->middleware(['adminrole']);
+Route::get('/createavatar', [AvatarController::class,'create'])->name('createavatar')->middleware(['adminrole']);
+Route::post('/storeavatar', [AvatarController::class,'store'])->middleware(['adminrole']);
+Route::get('/showavatar', [AvatarController::class,'show'])->name('showavatar')->middleware(['adminrole']);
+Route::get('/editavatar/{id}', [AvatarController::class,'edit'])->name('editavatar')->middleware(['adminrole']);
+Route::put('/{id}/updateavatar', [AvatarController::class,'update'])->middleware(['adminrole']);
+Route::delete('/{id}/deleteavatar', [AvatarController::class,'destroy'])->middleware(['adminrole']);
 
 // photos
 Route::get('/photos', [PhotoController::class,'index'])->name('photos');
@@ -66,10 +69,15 @@ Route::delete('/photos', [PhotoController::class,'destroy']);
 
 //gallery
 Route::get('/gallery', [GalleryController::class,'index'])->name('gallery');
+Route::post('/downloadpic/{id}', [GalleryController::class,'telecharge']);
 
 //category
-Route::get('/category', [CategoryController::class,'index'])->name('category');
-Route::delete('/{id}/deletecategory',[CategoryController::class,'destroy']);
+Route::get('/category', [CategoryController::class,'index'])->middleware(['adminrole'])->name('category');
+Route::get('/createcategory', [CategoryController::class,'create'])->middleware(['adminrole'])->name('createcategory');
+Route::post('/storecategory', [CategoryController::class,'store'])->middleware(['adminrole']);
+Route::get('/editcategory/{id}', [CategoryController::class,'edit'])->middleware(['adminrole'])->name('editcategory');
+Route::put('/{id}/updatecategory', [CategoryController::class,'update'])->middleware(['adminrole']);
+Route::delete('/{id}/deletecategory',[CategoryController::class,'destroy'])->middleware(['adminrole']);
 
 require __DIR__.'/auth.php';
